+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
 * Keychain storage that relies on UICKeyChainStore plug-in, only supports iOS
+* Add iOS Swift Package Manager support.